Rename fs module to platform
