+++ src/lib.rs
@@ -15,10 +15,10 @@ pub mod logger;
 
 mod camera;
 mod convert;
-mod fs;
 mod geometry;
 mod input;
 mod math;
+mod platform;
 mod renderer;
 mod ui;
